Support arbitrary functions for VolEsti integrator
