+++ .gitignore
@@ -1,3 +1,3 @@
 *.pyc
 *.py~
-experiments/srn/data
+experiments/**/*.json
