+++ wmipa/integration/integrand.py
@@ -1,0 +1,3 @@
+class Integrand:
+    def __init__(self):
+        self.variables = set()
