+++ wmipa/integration/polytope.py
@@ -7,293 +7,12 @@ is more suitable to be processed by LattE Integrale.
 __version__ = "0.99"
 __author__ = "Paolo Morettin"
 
-from fractions import Fraction
-
-import networkx as nx
-from pysmt.shortcuts import LT, And, Bool, Plus, Pow, Real, Symbol, Times
+from pysmt.shortcuts import LT, And, Bool, Plus, Real, Symbol, Times
 from pysmt.typing import REAL
 
-from wmipa.integration.sympy2pysmt import get_canonical_form
-from wmipa.utils import is_pow, lcmm
+from wmipa.integration.polynomial import Polynomial
+from wmipa.utils import lcmm
 from wmipa.wmiexception import WMIParsingException
-
-
-class Monomial:
-    """Intermediate representation of a monomial.
-
-    c * x_1^e_1 * ... * x_k^e_k
-
-    Attributes:
-        coefficient (Fraction): The coefficient of the monomial.
-        exponents (dict {FNode : Fraction}): The dictionary containing the
-            exponents {x_i : e_i}
-
-    """
-
-    def __init__(self, expression):
-        """Default constructor.
-
-        Takes as input a pysmt formula representing a monomial.
-
-        Args:
-            expression (FNode): The pysmt formula representing a monomial.
-
-        Raises:
-            WMIParsingException: If the expression is not a monomial.
-
-        """
-        if not (
-            expression.is_times()
-            or expression.is_real_constant()
-            or expression.is_symbol()
-            or is_pow(expression)
-        ):
-            raise WMIParsingException(WMIParsingException.NOT_A_MONOMIAL, expression)
-
-        self.coefficient, self.exponents = self._parse_sub(expression)
-
-    def __str__(self):
-        """The str method.
-
-        Returns:
-            str: The string representation of the monomial (e.g: (-6 * x^3 * y^2) ).
-
-        """
-        powers = ["{}^{}".format(k, e) for k, e in self.exponents.items()]
-        if len(self.exponents) == 0:
-            return "(" + str(self.coefficient) + ")"
-        else:
-            return "(" + str(self.coefficient) + " * " + " * ".join(powers) + ")"
-
-    def degree(self):
-        """Calculates the degree of the monomial.
-
-        This method evaluates the degree by summing up all the values of the exponents.
-
-        Returns:
-            int: The degree of the monomial.
-
-        """
-        return sum(self.exponents.values())
-
-    def multiply_by_monomial(self, monomial):
-        """Multiplies this instance with another monomial.
-
-        Args:
-            monomial (FNode): The other monomial.
-
-        """
-        assert isinstance(
-            monomial, Monomial
-        ), "Argument should be an instance of Monomial"
-
-        # multiply the coefficients
-        self.coefficient *= monomial.coefficient
-
-        # update all the exponents
-        for name, exp in monomial.exponents.items():
-            if name not in self.exponents:
-                self.exponents[name] = Fraction(0)
-            self.exponents[name] += exp
-
-    def negate(self):
-        """Negates the monomial by changing the sign of the coefficient."""
-        self.coefficient *= -1
-
-    def _parse_sub(self, expression):
-        """Parse an expression representing a monomial.
-
-        Args:
-            expression (FNode): The pysmt formula representing a monomial.
-
-        Returns:
-            Fraction: The coefficient of the expression.
-            (dict {FNode : Fraction}): The dictionary containing the exponents.
-
-        Raises:
-            WMIParsingException: If the expression is not a monomial.
-
-        """
-
-        if expression.is_real_constant():
-            return expression.constant_value(), {}
-        elif expression.is_symbol():
-            return Fraction(1), {expression.symbol_name(): Fraction(1)}
-        elif is_pow(expression):
-            base, exp = expression.args()
-
-            # Check that the exponent is constant
-            if not exp.is_real_constant():
-                raise WMIParsingException(
-                    WMIParsingException.NOT_A_MONOMIAL, expression
-                )
-            exp_value = exp.constant_value()
-
-            # Parse the base
-            base_value, base_var = self._parse_sub(base)
-
-            ret_value = pow(base_value, exp_value)
-            for x in base_var.keys():
-                base_var[x] *= exp_value
-
-            return ret_value, base_var
-        elif expression.is_times():
-            ret_value = Fraction(1)
-            ret_var = {}
-
-            # Parse all the terms of the multiplication
-            for sub in expression.args():
-                sub_value, sub_var = self._parse_sub(sub)
-                ret_value *= sub_value
-
-                for x, v in sub_var.items():
-                    if x not in ret_var:
-                        ret_var[x] = 0
-                    ret_var[x] += v
-            return ret_value, ret_var
-        else:
-            raise WMIParsingException(WMIParsingException.NOT_A_MONOMIAL, expression)
-
-    def to_pysmt(self):
-        if not self.exponents:
-            return Real(self.coefficient)
-        return Times(
-            Real(self.coefficient),
-            *map(
-                lambda i: Pow(Symbol(i[0], REAL), Real(i[1])),
-                self.exponents.items(),
-            ),
-        )
-
-
-class Polynomial:
-    """Intermediate representation of a polynomial.
-
-    Attributes:
-        monomials (Monomial): The list of Monomial instances.
-        variables (set(FNode)): The set of all the variables of the polynomial.
-
-    """
-
-    def __init__(self, expression, aliases={}):
-        """Default constructor.
-
-        Takes as input a pysmt formula representing a polynomial and a dict of
-        aliases to be substituted before the parsing.
-
-        Args:
-            expression (FNode): The pysmt formula representing the polynomial.
-            aliases (dict {FNode : FNode}): The dict containing the aliases definitions.
-
-        """
-        # Perform aliases substitution and put in canonical form
-        canonical = Polynomial._preprocess_formula(expression, aliases)
-
-        # Process every monomial individually
-        self.monomials = []
-        self.variables = set()
-        if canonical.is_plus():
-            for term in canonical.args():
-                self._add_monomial(Monomial(term))
-        else:
-            self._add_monomial(Monomial(canonical))
-
-    def __str__(self):
-        """The str method.
-
-        Returns:
-            str: The string representation of the polynomial
-                (e.g: (3 * x^5) + (-2 * y^3)).
-
-        """
-        return " + ".join(map(str, self.monomials))
-
-    def degree(self):
-        """Calculates the degree of the polynomial.
-
-        This method evaluates the degree by taking the maximum degree of its monomials.
-
-        Returns:
-            int: The degree of the polynomial.
-
-        """
-        if len(self.monomials) > 0:
-            return max([m.degree() for m in self.monomials])
-        else:
-            return 0
-
-    def negate(self):
-        """Negates the polinomial by negating all its monomials."""
-        for monomial in self.monomials:
-            monomial.negate()
-
-    def _add_monomial(self, monomial):
-        """Add a monomial to the polynomial.
-
-        Args:
-            monomial (Monomial): The monomial to be added.
-
-        """
-        assert isinstance(
-            monomial, Monomial
-        ), "Argument should be an instance of Monomial"
-        if monomial.coefficient != 0:
-            self.monomials.append(monomial)
-            self.variables = self.variables.union(set(monomial.exponents.keys()))
-
-    @staticmethod
-    def _preprocess_formula(expression, aliases):
-        """Preprocesses a pysmt polynomial by substituting the aliases and
-        rewriting it in canonical form.
-
-        Args:
-            expression (FNode): The pysmt formula representing a polynomial.
-            aliases (dict {FNode : FNode}): The dict containing the aliases definitions.
-
-        Returns:
-            FNode: The pysmt formula in canonical form with all the aliases substituted.
-
-        Raises:
-            WMIParsingException: If in the aliases there is a cyclic assignment
-                (e.g: x = y, y = x).
-        """
-        # Apply alias substitutions in the correct order
-        if len(aliases) > 0:
-            # Build a dependency graph of the substitutions and apply them in
-            # topological order
-            Gsub = nx.DiGraph()
-            constant_subs = {}
-
-            # For every alias
-            for x, alias_expr in aliases.items():
-                for y in alias_expr.get_free_variables():
-                    # Create a node from the alias to every symbol inside it
-                    Gsub.add_edge(x, y)
-                # If the alias substitution leads to a constant value (e.g: PI = 3.1415)
-                if len(alias_expr.get_free_variables()) == 0:
-                    constant_subs.update({x: alias_expr})
-
-            # Get the nodes in topological order
-            try:
-                sorted_substitutions = [
-                    node for node in nx.topological_sort(Gsub) if node in aliases
-                ]
-            except nx.exception.NetworkXUnfeasible:
-                raise WMIParsingException(
-                    WMIParsingException.CYCLIC_ASSIGNMENT_IN_ALIASES, aliases
-                )
-
-            # Apply all the substitutions
-            for alias in sorted_substitutions:
-                expression = expression.substitute({alias: aliases[alias]})
-            expression = expression.substitute(constant_subs)
-
-        return get_canonical_form(expression)
-
-    def to_pysmt(self):
-        if not self.monomials:
-            return Real(0)
-        return Plus(*map(lambda x: x.to_pysmt(), self.monomials))
 
 
 class Bound:
@@ -390,8 +109,8 @@ class Bound:
             lcd = lcmm(denominators)
             for monomial in poly.monomials:
                 assert (
-                    len(monomial.exponents) == 1
-                    and list(monomial.exponents.values())[0] == 1
+                        len(monomial.exponents) == 1
+                        and list(monomial.exponents.values())[0] == 1
                 ), "Not a monomial of degree 1"
                 name = list(monomial.exponents.keys())[0]
                 assert name not in self.coefficients, "Polynomial not in canonical form"
@@ -419,8 +138,8 @@ class Bound:
     def __eq__(self, other):
         if isinstance(other, Bound):
             return (
-                self.constant == other.constant
-                and self.coefficients == other.coefficients
+                    self.constant == other.constant
+                    and self.coefficients == other.coefficients
             )
         return False
